moved cmp config to lsp folder
