+++ init.lua
@@ -51,4 +51,3 @@ vim.api.nvim_create_autocmd({'BufEnter', 'QuitPre'}, {
     end
   end
 })
-
