+++ lua/modules/functions/config.lua
@@ -3,7 +3,6 @@ local config = {}
 function config.ufo()
    local ufo = require("ufo")
 
-
    vim.o.foldcolumn = '1' -- '0' is not bad
    vim.o.foldlevel = 99 -- Using ufo provider need a large value, feel free to decrease the value
    vim.o.foldlevelstart = 99
@@ -30,9 +29,6 @@ function config.ufo()
    ufo.setup{}
 end
 
-function config.cmp()
-   --
-end
 
 function config.treesitter()
    require("nvim-treesitter").setup{
