+++ lua/modules/functions/package.lua
@@ -14,16 +14,3 @@ package({ "kevinhwang91/nvim-ufo",
    dependencies = 'kevinhwang91/promise-async',
    config = config.ufo,
 })
-
--- Autocompletion/snippets/and stuff
-package({
-   "hrsh7th/nvim-cmp",
-   dependencies = {
-      "hrsh7th/cmp-nvim-lsp",
-      "hrsh7th/cmp-buffer",
-      "hrsh7th/cmp-path",
-      "hrsh7th/cmp-cmdline"
-   },
-   config = config.cmp,
-})
-
