+++ lua/modules/lsp/config.lua
@@ -4,8 +4,43 @@ function config.mason()
   require('mason').setup{}
 end
 
-function config.lsp()
-  require'lspconfig'.lua_ls.setup{}
+function config.cmp()
+   local cmp = require'cmp'
+
+   cmp.setup({
+      snippet = {
+         -- REQUIRED - you must specify a snippet engine
+         expand = function(args)
+            vim.fn["vsnip"](args.body) -- For `vsnip` users.
+            -- require('luasnip').lsp_expand(args.body) -- For `luasnip` users.
+            -- require('snippy').expand_snippet(args.body) -- For `snippy` users.
+            -- vim.fn["UltiSnips#Anon"](args.body) -- For `ultisnips` users.
+            -- vim.snippet.expand(args.body) -- For native neovim snippets (Neovim v0.10+)
+         end,
+      },
+      window = {
+         -- completion = cmp.config.window.bordered(),
+         -- documentation = cmp.config.window.bordered(),
+      },
+      mapping = cmp.mapping.preset.insert({
+         ['<C-b>'] = cmp.mapping.scroll_docs(-4),
+         ['<C-f>'] = cmp.mapping.scroll_docs(4),
+         ['<C-Space>'] = cmp.mapping.complete(),
+         ['<C-e>'] = cmp.mapping.abort(),
+         ['<CR>'] = cmp.mapping.confirm({ select = true }), -- Accept currently selected item. Set `select` to `false` to only confirm explicitly selected items.
+      }),
+      sources = cmp.config.sources({
+         { name = 'nvim_lsp' },
+         { name = 'vsnip' }, -- For vsnip users.
+         -- { name = 'luasnip' }, -- For luasnip users.
+         -- { name = 'ultisnips' }, -- For ultisnips users.
+         -- { name = 'snippy' }, -- For snippy users.
+      }, {
+         { name = 'buffer' },
+      })
+   })
+
+   require'lspconfig'.lua_ls.setup{}
 end
 
 return config
