+++ lua/modules/lsp/package.lua
@@ -8,8 +8,18 @@ package({
    config = config.mason,
 })
 
-
--- Nvim LSP configs
-package({ "neovim/nvim-lspconfig",
-   config = config.lsp,
+-- Lsp and autocompletion
+package({
+   "hrsh7th/nvim-cmp",
+   dependencies = {
+      "hrsh7th/vim-vsnip",
+      --"hrsh7th/vim-vsnip-integ",
+      "hrsh7th/cmp-vsnip",
+      "hrsh7th/cmp-nvim-lsp",
+      "hrsh7th/cmp-buffer",
+      "hrsh7th/cmp-path",
+      "hrsh7th/cmp-cmdline",
+      "neovim/nvim-lspconfig"
+   },
+   config = config.cmp,
 })
